LPS-192541 Extract ImportLayoutsMVCActionCommand methods to ImportLayoutsMVCActionCommandHelper and apply to usages
